+++ src/imio/smartweb/core/rest/authentic_sources.py
@@ -48,6 +48,7 @@ class BaseRequestForwarder(Service):
         else:
             token = get_wca_token(self.client_id, self.client_secret)
             headers = {"Accept": "application/json", "Authorization": token}
+        self.request.form.pop("wcatoken", None)
         if is_log_active():
             logger.info("======== Forwarding request to AUTHENTIC SOURCE =========")
             logger.info(f"url to forward : {url} ({method})")